Tidy infinite phase screen module
