+++ aotools/turbulence/__init__.py
@@ -3,3 +3,4 @@ from .infinitephasescreen import *
 from .temporal_ps import *
 from .opticalpropagation import *
 from .slopecovariance import *
+from .turbulence import *
